Update task 2.py

+++ buns/mod2/task 2.py
@@ -1,4 +1,4 @@
-square_side = int(input("Введите длину стороны квадрата: "))
+square_side = float(input("Введите длину стороны квадрата: "))
 square_perimeter = square_side * 4
 square_area = square_side ** 2
 square_diagonal = round(square_side * (2 ** 0.5), 2)
